chore(ci): replace service-specific test workflows with a unified workflow (#56)
Closes #43 

Running tests against individual service packages introduces a risk
(however unlikely) of missed interactions between services and doesn't
cover potential future code that sits outside of service packages. This
removes the existing service-specific test workflows as well as the
template for those workflows.

The new workflow also lints all code to ensure it meets the configured
standards; since the workflow uses `golangci-lint-action` to perform the
linting, a new `make lint` task has been added so that it is easier to
reproduce any failures locally.

Note that the `make test` task already tested the full SDK, it just
didn't ensure that all code was up-to-date first. The `test` task is
part of the `generate` task in each service-specific Makefile; we should
probably remove it from there and have an explicit `make test` step in
the GitHub Actions workflow so that we aren't running tests multiple
times.