chore: explicitly remove temporary merged metal spec after codegen (#93)
The semantic-release tool explicitly ignores the `.gitignore` file,
which means that the automated release process will commit all files
that are present in the local repo even if they are in the `.gitignore`.
This updates the Metal codegen task to explicitly clean up the temporary
merged spec. If swagger-api/swagger-parser#2105 gets merged and is
included in an openapi-generator release before this SDK loses support
for Metal, we can remove the temporary merged Metal spec.