chore: fix path to temp Metal spec files (#95)
The `openapi-generator` commands run in a container, so paths are
prefixed with `/local`. The `rm` commands run on the host machine, so
paths are not prefixed with `/local`. We should fix this discrepancy to
avoid more issues with remembering which path to use for which command,
but for now this PR just fixes the `rm` command for the temporary merged
Metal spec so we stop reintroducing unwanted files on every release.