chore: stop generating skeleton tests (#100)
The tests produced by openapi-generator are understandably not very
useful. To make them run at all we would have to remove the `t.Skip`
calls. That only requires a custom template, but to make them effective
we would have to make some serious modifications that cannot be
templated, either by stringing multiple tests together for end-to-end
tests or by heavily mocking out the live APIs for isolated SDK-only
tests.

If we add usable & meaningful tests directly to this SDK in the future,
they should be placed outside of the `services` directory to avoid
unnecessary complication in our codegen logic. In the meantime, we can
leverage the Equinix Terraform provider tests to validate changes to
this SDK.