feat: upgrade openapi-generator to v7.12.0 (#99)
This upgrades the code generator to v7.12.0 and regenerates the SDK. The
effects of this upgrade are:

- We no longer need a custom `model_oneof` template for EIA because the
bug that necessitated that template has been fixed
- The `MarshalJSON` method for `anyOf` models now uses a value receiver
to match other marshal methods
- The `anyOf` models gain a `GetActualInstanceValue` method

The new version of openapi-generator also introduces a prefix on enum
values. To minimize the impact of this upgrade, the
`removeEnumValuePrefix` has been set to `true` in the generator config
file to disable the new generator behavior.