chore(ci): more consistent handling of failed spec patches (#110)
This updates the `patch` tasks and spec sync workflows so that we handle
failed patches more consistently. For the most part, patches fail when
the equivalent changes are adopted in the upstream spec, although there
are other potential failure modes.

Prior to this change, a `make patch` task would succeed as long as the
last applied patch succeeds. Failed patches in the middle of the task
would be silently ignored, but a failure of the _last_ patch would cause
the `make patch` task to fail, and thus cause `sync` workflows to fail
as well.

This changes the `patch` tasks so that _any_ failed patch will result in
a non-zero error code. To counterbalance this change, the `sync`
workflows are updated so that, even if a patch fails, the workflow will
continue with generating updated code. This means that `sync` jobs will
be able to open a PR with updated code even if a patch was adopted
upstream, making it easier for teams to pull spec updates into this SDK
and then clean up dangling `.rej` files afterwards.