feat: onboard accesstokenv1 service (#102)
This adds a service covering the `/oauth2/v1` API endpoints. One oddity
here is that the service is called `accesstokenv1`, which matches its
name according to the API catalog, but per the recommendation from `make
onboard-service`, it _should_ be called `oauth2v1`.

This service is used by equinix/terraform-provider-equinix#847. In the
Terraform provider we need to be able to reuse the same token across
multiple request contexts to avoid making unnecessary duplicate requests
to the token service. That means we need to introduce some custom types
based on `golang.org/x/oauth2` because [the `oauth2` library does not
accept a context at the time a token is
obtained](https://github.com/golang/oauth2/issues/262) (in the core
library, the context can only be specified when a token source is
instantiated).