fix(docs): update examples to use built-in OAuth2 extension (#122)
Support for the [access token
service](https://docs.equinix.com/api-catalog/accesstokenv1) was added
in `v0.53.0`, but we still had some examples using the abandoned
`equinix/oauth2-go` library that was created for Terraform usage. This
updates the examples to use the built-in access token integration so we
don't send users in the wrong direction.