chore: adopt openapi-generator-ignore file to control output (#127)
This replaces the `modelTests` and `apiTests` settings in the generator
config file with a custom openapi-generator-ignore file so that we can
more explicitly and thoroughly control what files are produced by the
generator in each service package. Switching to an ignore file
eliminates the need for the `remove-unused` cleanup task, since we can
now avoid generating unwanted files in the first place.