chore: stop regenerating go.mod (#128)
Previously, we regenerated `go.mod` every time the SDK was regenerated,
but doing so is unnecessary. This updates the `mod` task to run `go mod
tidy` without recreating `go.mod`. This also updates the GitHub Actions
workflows to install the Go version specified in `go.mod` rather than
hard-coding a version that may not match.